refactor: import media interfaces from @tinacms/core
